Modified and redesign code
