+++ FrontEnd/package.json
@@ -28,7 +28,8 @@
     "react-icons": "^5.1.0",
     "react-router-dom": "^6.18.0",
     "react-webcam": "^7.2.0",
-    "styled-components": "^6.1.0"
+    "styled-components": "^6.1.0",
+    "tw-elements-react": "^1.0.0-alpha2"
   },
   "devDependencies": {
     "@types/react": "^18.2.15",
