+++ FrontEnd/src/Components/SignupForm.jsx
@@ -1,98 +1,27 @@
-import Avatar from "@mui/material/Avatar";
-import Button from "@mui/material/Button";
-import CssBaseline from "@mui/material/CssBaseline";
-import TextField from "@mui/material/TextField";
-import Link from "@mui/material/Link";
-import Grid from "@mui/material/Grid";
-import Box from "@mui/material/Box";
-import Typography from "@mui/material/Typography";
-import Container from "@mui/material/Container";
-import { Link as RouterLink, useNavigate } from "react-router-dom";
-import { useContext, useState } from "react";
+import { useState, useContext } from "react";
 import axios from "axios";
-import { IconButton, InputAdornment } from "@mui/material";
-import VisibilityOffIcon from "@mui/icons-material/VisibilityOff";
-import VisibilityIcon from "@mui/icons-material/Visibility";
-import toast from "react-hot-toast";
 import UserContext from "./UserContext";
+import sidebarlogo from "../assets/Images/sidebarlogo.png";
+import { Link as RouterLink, useNavigate } from "react-router-dom";
 
-function Copyright(props) {
-  return (
-    <Typography
-      variant="body2"
-      color="text.secondary"
-      align="center"
-      {...props}
-    >
-      {"Made with 💖 by "}
-      <Link color="inherit" href="http://sandyapps.co">
-        SandyApps
-      </Link>
-      {" ©️ "}
-      {new Date().getFullYear()}
-      {"."}
-    </Typography>
-  );
-}
-
-// TODO remove, this demo shouldn't need to reset the theme.
-
-export default function SignupForm() {
-  const { Api_EndPoint } = useContext(UserContext);
+export default function SignupForm({ login, role }) {
   const [showPassword, setShowPassword] = useState(false);
-  const navigate = useNavigate();
-  const [formData, setFormData] = useState({
-    name: "",
-    userName: "",
-    email: "",
-    password: "",
-    confirmPassword: "",
-  });
-
+  const [formData, setFormData] = useState({ username: "", password: "" });
   const [errors, setErrors] = useState({});
-
-  const handleChanges = (event) => {
-    setFormData({ ...formData, [event.target.name]: event.target.value });
-  };
+  const { setUserName, setUserRole, Api_EndPoint } = useContext(UserContext);
+  const navigate = useNavigate();
 
   const validateForm = () => {
     let isValid = true;
     const newErrors = {};
 
-    // Validate email
-    const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
-    const passwordRegex =
-      /^(?=.*[A-Z])(?=.*[a-z])(?=.*[0-9])(?=.*[!@#$%^&*()_+=-{}|:;\\'\\",<.>/?\\`~]).{8,}$/;
-
-    if (!formData.name) {
-      newErrors.name = "Enter Name";
-      isValid = false;
-    }
-    if (!formData.userName) {
-      newErrors.name = "Enter a unique user Name";
-      isValid = false;
-    }
-    if (!formData.email || !emailRegex.test(formData.email)) {
-      newErrors.email = "Enter a valid email address";
+    if (!formData.username) {
+      newErrors.username = "Please enter username";
       isValid = false;
     }
 
-    // Validate password
-    if (
-      formData.password.length < 7 ||
-      formData.password.length > 15 ||
-      !passwordRegex.test(formData.password)
-    ) {
-      newErrors.password = "Enter a valid password between 7 and 15 letters";
-      toast.error(
-        "Password must contain at least 1 uppercase letter, 1 lowercase letter, 1 number and 1 special character"
-      );
-      isValid = false;
-    }
-
-    // Validate confirm password
-    if (formData.password !== formData.confirmPassword) {
-      newErrors.confirmPassword = "Passwords do not match";
+    if (formData.password.length < 7) {
+      newErrors.password = "Password must be at least 7 characters";
       isValid = false;
     }
 
@@ -100,191 +29,199 @@ export default function SignupForm() {
     return isValid;
   };
 
-  const handleSubmit = async (event) => {
-    event.preventDefault();
+  const handleSubmit = async (e) => {
+    e.preventDefault();
 
     if (validateForm()) {
       try {
-        // Check if the username already exists
-        const usernameExists = await axios.get(
-          `${Api_EndPoint}/api/users/exists/${formData.userName}`
-        );
-        if (usernameExists.data.exists) {
-          // Username already exists, display an error
-          toast.error(
-            "Username already exists. Please choose a different username."
-          );
-          return;
-        }
-        const apiEndpoint = "/api/users";
-        await axios.post(`${Api_EndPoint}${apiEndpoint}`, {
-          name: formData.name,
-          username: formData.userName,
-          email: formData.email,
+        const response = await axios.post(`${Api_EndPoint}/api/users/login`, {
+          username: formData.username,
           password: formData.password,
         });
-        toast.success("Registered Successfully");
-        navigate("/");
-        // Optionally, you can redirect the user or perform other actions after successful registration
+
+        setUserName(formData.username);
+        login(true);
+        role(response.data.role);
+        setUserRole(response.data.role);
+        navigate("/home");
       } catch (error) {
-        // Handle errors from the server
-        console.error("Error registering user:", error);
+        console.error("Login failed:", error);
+        // Handle login error here
+        setErrors({
+          username: "Invalid username or password",
+          password: "Invalid username or password",
+        });
       }
     } else {
-      // Form is not valid, handle accordingly
-      toast.error("Form is not valid");
+      console.log("Invalid Form");
     }
   };
 
+  const handleChange = (e) => {
+    const { name, value } = e.target;
+    setFormData({ ...formData, [name]: value });
+  };
+
   return (
-    <Container component="main" maxWidth="xs" sx={{ color: "text.primary" }}>
-      <CssBaseline />
-      <Box
-        sx={{
-          marginTop: 8,
-          display: "flex",
-          flexDirection: "column",
-          alignItems: "center",
-        }}
-      >
-        <Avatar sx={{ m: 1, bgcolor: "white" }}>
-          <img src="/src/assets/logo.png" width="28px" height="auto" />
-        </Avatar>
-        <Typography component="h1" variant="h5">
-          Sign Up
-        </Typography>
-        <Box component="form" onSubmit={handleSubmit} noValidate sx={{ mt: 1 }}>
-          <TextField
-            margin="normal"
-            required
-            fullWidth
-            id="name"
-            label="Name"
-            name="name"
-            autoComplete="text"
-            autoFocus
-            error={errors.name}
-            value={formData.name}
-            onChange={handleChanges}
-            inputProps={{
-              maxLength: 30, // Limiting to 30 characters
-            }}
+    <div className="min-h-screen flex">
+      <div className="hidden lg:flex w-full lg:w-1/2 login_img_section justify-around items-center bg-gradient-to-r from-sky-600 to-cyan-400 text-white">
+        <div className="bg-black opacity-20 inset-0 z-0"></div>
+        <div className="w-full mx-auto px-20 flex-col items-center space-y-6">
+          <p className="text-9xl text-white mt-1">Attendance</p>
+          <img
+            className="max-w-48 ml-auto"
+            src={sidebarlogo}
+            alt="logo"
+            // Adjust the height, width, and color as needed
           />
-          <TextField
-            margin="normal"
-            required
-            fullWidth
-            id="userName"
-            label="User Name"
-            name="userName"
-            autoComplete="text"
-            error={errors.userName}
-            value={formData.userName}
-            onChange={handleChanges}
-            inputProps={{
-              maxLength: 25, // Limiting to 25 characters
-            }}
-          />
-          <TextField
-            margin="normal"
-            required
-            fullWidth
-            id="email"
-            label="Email Address"
-            name="email"
-            autoComplete="email"
-            value={formData.email}
-            onChange={handleChanges}
-            error={!!errors.email}
-            helperText={errors.email}
-            inputProps={{
-              maxLength: 35, // Limiting to 35 characters
-            }}
-          />
-          <TextField
-            margin="normal"
-            required
-            fullWidth
-            name="password"
-            label="Password"
-            type={showPassword ? "text" : "password"}
-            id="password"
-            autoComplete="current-password"
-            value={formData.password}
-            onChange={handleChanges}
-            error={errors.password}
-            helperText={errors.password}
-            InputProps={{
-              endAdornment: (
-                <InputAdornment position="end">
-                  <IconButton
-                    aria-label="toggle password visibility"
-                    onClick={() => setShowPassword(!showPassword)}
-                    edge="end"
-                  >
-                    {showPassword ? <VisibilityOffIcon /> : <VisibilityIcon />}
-                  </IconButton>
-                </InputAdornment>
-              ),
-            }}
-            inputProps={{
-              maxLength: 20, // Limiting to 20 characters
-            }}
-          />
-          <TextField
-            margin="normal"
-            required
-            fullWidth
-            name="confirmPassword"
-            label="Confirm-Password"
-            type={showPassword ? "text" : "password"}
-            id="confirmPassword"
-            autoComplete="confirm-password"
-            value={formData.confirmPassword}
-            onChange={handleChanges}
-            error={!!errors.confirmPassword}
-            helperText={errors.confirmPassword}
-            InputProps={{
-              endAdornment: (
-                <InputAdornment position="end">
-                  <IconButton
-                    aria-label="toggle password visibility"
-                    onClick={() => setShowPassword(!showPassword)}
-                    edge="end"
-                  >
-                    {showPassword ? <VisibilityOffIcon /> : <VisibilityIcon />}
-                  </IconButton>
-                </InputAdornment>
-              ),
-            }}
-            inputProps={{
-              maxLength: 20, // Limiting to 20 characters
-            }}
-          />
-          <Button
-            type="submit"
-            fullWidth
-            variant="contained"
-            sx={{
-              mt: 3,
-              mb: 2,
-              bgcolor: "#1db0e6",
-              "&:hover": { bgcolor: "#1b1d72" },
-            }}
+          <p className="text-4xl text-white mt-1">Welcome to Daily TimeSheet</p>
+          <div className="flex justify-end  mt-6">
+            <a
+              href="#"
+              className="hover:bg-gradient-to-r from-cyan-400 to-sky-600 hover:text-white hover:-translate-y-1 transition-all duration-500 bg-white text-cyan-700 mt-4 px-4 py-2 rounded-2xl font-bold mb-2"
+            >
+              Get Started
+            </a>
+          </div>
+        </div>
+      </div>
+      <div className="flex w-full lg:w-1/2 justify-center items-center bg-white space-y-8 mt-10 p-5">
+        <div className="w-full px-8 md:px-32 lg:px-24">
+          <form
+            className="bg-[#DBF3FA] rounded-md shadow-2xl p-5 mb-10"
+            onSubmit={handleSubmit}
           >
-            Sign up
-          </Button>
-          <Grid container justifyContent="flex-end">
-            <Grid item></Grid>
-            <Grid item>
-              <Link component={RouterLink} to="/" variant="body2">
-                {"Already have an account? login"}
-              </Link>
-            </Grid>
-          </Grid>
-        </Box>
-      </Box>
-      <Copyright sx={{ mt: 8, mb: 4 }} />
-    </Container>
+            <h1 className="text-gray-800 font-bold text-2xl mb-10 flex flex-col justify-center  ">
+              Signup
+            </h1>
+            <div className="flex items-center  mb-8 py-2 px-3 rounded-2xl">
+              <svg
+                xmlns="http://www.w3.org/2000/svg"
+                className="h-5 w-5 text-black-400"
+                fill="none"
+                viewBox="0 0 24 24"
+                stroke="currentColor"
+              >
+                <path
+                  strokeLinecap="round"
+                  strokeLinejoin="round"
+                  strokeWidth="2"
+                  d="M16 12a4 4 0 10-8 0 4 4 0 008 0zm0 0v1.5a2.5 2.5 0 005 0V12a9 9 0 10-9 9m4.5-1.206a8.959 8.959 0 01-4.5 1.207"
+                />
+              </svg>
+              &nbsp;
+              <input
+                id="username"
+                className={`pl-2 w-full outline-none border border-gray-300 rounded-md py-2 ${
+                  errors.username ? "border-red-500" : ""
+                }`}
+                type="text"
+                name="username"
+                placeholder="Username"
+                value={formData.username}
+                onChange={handleChange}
+              />
+            </div>
+            {errors.username && (
+              <p className="text-red-500 text-xs">{errors.username}</p>
+            )}
+            <div className="flex items-center mb-12 py-2 px-3 rounded-2xl">
+              <svg
+                xmlns="http://www.w3.org/2000/svg"
+                className="h-5 w-5 text-gray-400"
+                viewBox="0 0 20 20"
+                fill="currentColor"
+              >
+                <path
+                  fillRule="evenodd"
+                  d="M5 9V7a5 5 0 0110 0v2a2 2 0 012 2v5a2 2 0 01-2 2H5a2 2 0 01-2-2v-5a2 2 0 012-2zm8-2v2H7V7a3 3 0 016 0z"
+                  clipRule="evenodd"
+                />
+              </svg>
+              &nbsp;
+              <input
+                className={`pl-2 w-full outline-none border border-gray-300 rounded-md py-2 ${
+                  errors.password ? "border-red-500" : ""
+                }`}
+                type={showPassword ? "text" : "password"}
+                name="password"
+                id="password"
+                placeholder="Password"
+                value={formData.password}
+                onChange={handleChange}
+              />
+              &nbsp;
+              <button
+                type="button"
+                onClick={() => setShowPassword(!showPassword)}
+              >
+                {showPassword ? (
+                  <svg
+                    xmlns="http://www.w3.org/2000/svg"
+                    className="h-5 w-5 text-gray-400 cursor-pointer"
+                    viewBox="0 0 24 24"
+                    fill="none"
+                    stroke="currentColor"
+                  >
+                    <path
+                      strokeLinecap="round"
+                      strokeLinejoin="round"
+                      strokeWidth="2"
+                      d="M6 18L18 6M6 6l12 12"
+                    />
+                  </svg>
+                ) : (
+                  <svg
+                    xmlns="http://www.w3.org/2000/svg"
+                    className="h-5 w-5 text-gray-400 cursor-pointer"
+                    viewBox="0 0 24 24"
+                    fill="none"
+                    stroke="currentColor"
+                  >
+                    <path
+                      strokeLinecap="round"
+                      strokeLinejoin="round"
+                      strokeWidth="2"
+                      d="M15 12a3 3 0 11-6 0 3 3 0 016 0z"
+                    />
+                    <path
+                      strokeLinecap="round"
+                      strokeLinejoin="round"
+                      strokeWidth="2"
+                      d="M7 7a3 3 0 00-3 3v2a3 3 0 003 3h10a3 3 0 003-3v-2a3 3 0 00-3-3"
+                    />
+                  </svg>
+                )}
+              </button>
+            </div>
+            {errors.password && (
+              <p className="text-red-500 text-xs">{errors.password}</p>
+            )}
+            <div className="flex justify-center">
+              <button
+                type="submit"
+                className="mx-auto hover:bg-gradient-to-r from-cyan-400 to-sky-600 hover:text-white hover:-translate-y-1 transition-all duration-500 bg-white text-cyan-700 mt-4 px-28 py-2 rounded-2xl font-bold mb-2"
+              >
+                Login
+              </button>
+            </div>
+            <div className="flex justify-between mt-4">
+              <RouterLink
+                to="#"
+                className="text-sm ml-2 hover:text-cyan-600 cursor-pointer hover:-translate-y-1 duration-500 transition-all"
+              >
+                Forgot Password?
+              </RouterLink>
+              <RouterLink
+                to="/signup"
+                className="text-sm ml-2 hover:text-cyan-600 cursor-pointer hover:-translate-y-1 duration-500 transition-all"
+              >
+                Don't have an account yet?
+              </RouterLink>
+            </div>
+          </form>
+        </div>
+      </div>
+    </div>
   );
 }
